+++ org-roam-graph.el
@@ -67,6 +67,20 @@ Example:
   :type '(alist)
   :group 'org-roam)
 
+(defcustom org-roam-graph-edge-extra-config nil
+  "Extra options for graphviz edges.
+Example:
+ '((\"dir\" . \"back\"))"
+  :type '(alist)
+  :group 'org-roam)
+
+(defcustom org-roam-graph-edge-cites-extra-config '(("color" . "red"))
+  "Extra options for graphviz edges for citation links.
+Example:
+ '((\"dir\" . \"back\"))"
+  :type '(alist)
+  :group 'org-roam)
+
 (defcustom org-roam-graph-max-title-length 100
   "Maximum length of titles in graph nodes."
   :type 'number
@@ -132,24 +146,41 @@ into a digraph."
             `[:with selected :as [:select [file] :from ,node-query]
               :select :distinct [to from] :from links
               :where (and (in to selected) (in from selected))])
-           (edges (org-roam-db-query edges-query)))
+           (edges-cites-query
+            `[:with selected :as [:select [file] :from ,node-query]
+              :select :distinct [file from]
+              :from links :inner :join refs :on (and (= links:to refs:ref) (= links:type "cite"))
+              :where (and (in file selected) (in from selected))])
+           (edges (org-roam-db-query edges-query))
+           (edges-cites (org-roam-db-query edges-cites-query)))
       (insert "digraph \"org-roam\" {\n")
+
       (dolist (option org-roam-graph-extra-config)
-        (insert (concat (car option)
+        (insert (concat "  "
+                        (car option)
                         "="
                         (cdr option)
                         ";\n")))
+      (insert (format "  node [%s];\n"
+                      (->> org-roam-graph-node-extra-config
+                           (mapcar (lambda (n)
+                                     (concat (car n) "=" (cdr n))))
+                           (s-join ","))))
+      (insert (format "  edge [%s];\n"
+                      (->> org-roam-graph-edge-extra-config
+                           (mapcar (lambda (n)
+                                     (concat (car n) "=" (cdr n))))
+                           (s-join ","))))
+
       (dolist (node nodes)
         (let* ((file (xml-escape-string (car node)))
                (title (or (caadr node)
                           (org-roam--path-to-slug file)))
                (shortened-title (s-truncate org-roam-graph-max-title-length title))
-               (base-node-properties (list (cons "label" (s-replace "\"" "\\\"" shortened-title))
-                                           (cons "URL" (concat "org-protocol://roam-file?file="
-                                                               (url-hexify-string file)))
-                                           (cons "tooltip" (xml-escape-string title))))
-               (node-properties (append base-node-properties
-                                        org-roam-graph-node-extra-config)))
+               (node-properties (list (cons "label" (s-replace "\"" "\\\"" shortened-title))
+                                      (cons "URL" (concat "org-protocol://roam-file?file="
+                                                          (url-hexify-string file)))
+                                      (cons "tooltip" (xml-escape-string title)))))
           (insert
            (format "  \"%s\" [%s];\n"
                    file
@@ -161,14 +192,23 @@ into a digraph."
         (insert (format "  \"%s\" -> \"%s\";\n"
                         (xml-escape-string (car edge))
                         (xml-escape-string (cadr edge)))))
+
+      (insert (format "  edge [%s];\n"
+                      (->> org-roam-graph-edge-cites-extra-config
+                           (mapcar (lambda (n)
+                                     (concat (car n) "=" (cdr n))))
+                           (s-join ","))))
+      (dolist (edge edges-cites)
+        (insert (format "  \"%s\" -> \"%s\";\n"
+                        (xml-escape-string (car edge))
+                        (xml-escape-string (cadr edge)))))
       (insert "}")
       (buffer-string))))
 
-(defun org-roam-graph-show (&optional prefix node-query)
-  "Generate and displays the Org-roam graph using `org-roam-graph-viewer'.
-If PREFIX, then the graph is generated but the viewer is not invoked."
-  (interactive "P")
-  (declare (indent 0))
+(defun org-roam-graph-build (&optional node-query)
+  "Generate a graph showing the relations between nodes in NODE-QUERY.
+For building and showing the graph in a single step see `org-roam-graph-show'."
+  (interactive)
   (unless org-roam-graph-executable
     (user-error "Can't find %s executable.  Please check if it is in your path"
                 org-roam-graph-executable))
@@ -181,15 +221,25 @@ If PREFIX, then the graph is generated but the viewer is not invoked."
     (with-temp-file temp-dot
       (insert graph))
     (call-process org-roam-graph-executable nil 0 nil temp-dot "-Tsvg" "-o" temp-graph)
-    (unless prefix
-      (if (and org-roam-graph-viewer (executable-find org-roam-graph-viewer))
-	        (call-process org-roam-graph-viewer nil 0 nil temp-graph)
-        (view-file temp-graph)))))
+    temp-graph))
 
-(defun org-roam-graph-show-connected-component (&optional max-distance no-display)
-  "Like `org-roam-graph-show', but only show nodes connected to the current entry.
-If MAX-DISTANCE is non-nil, only nodes within the given number of steps are shown.
-If NO-DISPLAY is non-nil, the graph is generated but the viewer is not invoked."
+(defun org-roam-graph--open (file)
+  "Open FILE using `org-roam-graph-viewer', with `view-file' as a fallback."
+  (if (and org-roam-graph-viewer (executable-find org-roam-graph-viewer))
+      (call-process org-roam-graph-viewer nil 0 nil file)
+    (view-file temp-graph)))
+
+(defun org-roam-graph-show (&optional node-query)
+  "Generate and display a graph showing the relations between nodes in NODE-QUERY.
+The graph is generated using `org-roam-graph-build' and subsequently displayed
+using `org-roam-graph-viewer', if it refers to a valid executable, or using
+`view-file' otherwise."
+  (interactive)
+  (org-roam-graph--open (org-roam-graph-build node-query)))
+
+(defun org-roam-graph-build-connected-component (&optional max-distance)
+  "Like `org-roam-graph-build', but only include nodes connected to the current entry.
+If MAX-DISTANCE is non-nil, only nodes within the given number of steps are shown."
   (interactive "P")
   (unless (org-roam--org-roam-file-p)
     (user-error "Not in an Org-roam file"))
@@ -201,7 +251,13 @@ If NO-DISPLAY is non-nil, the graph is generated but the viewer is not invoked."
          (query `[:select [file titles]
                   :from titles
                   :where (in file [,@files])]))
-    (org-roam-graph-show no-display query)))
+    (org-roam-graph-build query)))
+
+(defun org-roam-graph-show-connected-component (&optional max-distance)
+  "Like `org-roam-graph-show', but only include nodes connected to the current entry.
+If MAX-DISTANCE is non-nil, only nodes within the given number of steps are shown."
+  (interactive "P")
+  (org-roam-graph--open (org-roam-graph-build-connected-component max-distance)))
 
 (provide 'org-roam-graph)